Se ajusta estructura del proyecto
Se ajusta la estructura del proyecto de acuerdo con lo expuesto en week3day2 moviendo y generando carpetas, también se ajusta el customPipe creado al igual que el templete usado para implementar otros pipes de Angular.
