--- src/app/index.ts
@@ -1,8 +1,0 @@
-export * from './app-routing.module';
-
-
-
-export * from './app.component';
-export * from './app.module';
-export * from './modules';
-export * from './shared';
