+++ src/app/modules/index.ts
@@ -1,3 +1,1 @@
-export * from './components';
-export * from './interfaces';
-export * from './services';
+export * from './main';
