+++ src/app/modules/main/index.ts
@@ -1,3 +1,5 @@
+export * from './app-routing.module';
+export * from './app.module';
 export * from './components';
 export * from './directives';
 export * from './enums';
