+++ src/app/modules/main/pages/app/app.component.ts
