+++ src/app/modules/main/pages/app/index.ts
@@ -1,0 +1,4 @@
+
+
+
+export * from './app.component';
