+++ src/app/modules/main/pages/auth/auth.component.html
