+++ src/app/modules/main/pages/auth/auth.component.spec.ts
