+++ src/app/modules/main/pages/auth/index.ts
