+++ src/app/modules/main/pages/index.ts
@@ -1,4 +1,5 @@
 export * from './accounts';
+export * from './app';
 export * from './auth';
 export * from './customers';
 export * from './dashboard';
